# modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        Node:
            description:
                https://www.schema.org/Node
            properties:
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                id:
                    format: int64
                    type: integer
                initial:
                    type: boolean
                name:
                    type: string
                processId:
                    format: int64
                    type: integer
                processVersion:
                    type: string
                terminal:
                    type: boolean
                type:
                    type: string
            type: object
...
        Task:
            description:
                https://www.schema.org/Task
            properties:
                assigneeId:
                    format: int64
                    type: integer
                className:
                    type: string
                classPK:
                    format: int64
                    type: integer
                completed:
                    type: boolean
                completionUserId:
                    format: int64
                    type: integer
                dateCompletion:
                    format: date-time
                    type: string
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                description:
                    type: string
                duration:
                    format: int64
                    type: integer
                id:
                    format: int64
                    type: integer
                instanceId:
                    format: int64
                    type: integer
                label:
                    readOnly: true
                    type: string
                name:
                    type: string
                processId:
                    format: int64
                    type: integer
                processVersion:
                    type: string
                tokenId:
                    format: int64
                    type: integer
            type: object
...
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
...
    "/processes/{processId}/nodes":
        get:
            operationId: getProcessNodesPage
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Node"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Node"
                                type: array
                    description:
                        ""
            tags: ["Node"]
        post:
            operationId: postProcessNode
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Node"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Node"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Node"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Node"
                    description:
                        ""
            tags: ["Node"]
    "/processes/{processId}/{processVersion}/nodes/{nodeId}":
        delete:
            operationId: deleteProcessNode
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: processVersion
                  required: true
                  schema:
                      type: string
                - in: path
                  name: nodeId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["Node"]
...
    "/processes/{processId}/tasks":
        get:
            operationId: getProcessTasksPage
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Task"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Task"
                                type: array
                    description:
                        ""
            tags: ["Task"]
        post:
            operationId: postProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Task"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Task"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Task"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Task"
                    description:
                        ""
            tags: ["Task"]
...
    "/processes/{processId}/tasks/{taskId}":
        delete:
            operationId: deleteProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: taskId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["Task"]
        get:
            operationId: getProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: taskId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Task"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Task"
                    description:
                        ""
            tags: ["Task"]
        put:
            operationId: putProcessTask
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: taskId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Task"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Task"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Task"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Task"
                    description:
                        ""
            tags: ["Task"]
...